+++ delve-derive/src/attributes/variant.rs
@@ -1,5 +1,5 @@
 use deluxe::{ParseAttributes, ParseMetaItem};
-use syn::{ExprClosure, Ident, LitStr, Token};
+use syn::{ExprClosure, Ident, LitStr, Path, Token};
 
 #[derive(Debug, ParseAttributes)]
 #[deluxe(attributes(delve))]
@@ -25,7 +25,7 @@ pub(crate) struct VariantAttribute {
 
 #[derive(Debug)]
 pub(crate) enum DisplayValue {
-    ExternFn(Ident),
+    ExternFn(Path),
     String(String),
 
     Closure(ExprClosure),
@@ -38,8 +38,8 @@ impl ParseMetaItem for DisplayValue {
     ) -> deluxe::Result<Self> {
         if input.peek(Token![|]) {
             Ok(DisplayValue::Closure(input.parse::<ExprClosure>()?))
-        } else if input.peek(Ident) {
-            Ok(DisplayValue::ExternFn(input.parse::<Ident>()?))
+        } else if input.peek(Ident) || input.peek(Token![::]) {
+            Ok(DisplayValue::ExternFn(input.parse::<Path>()?))
         } else {
             Ok(DisplayValue::String(input.parse::<LitStr>()?.value()))
         }